feat: Nav logic for Config pages ophuscated and unified into mixins
