+++ src/claudefig/tui/base/__init__.py
@@ -1,10 +1,11 @@
 """Base classes and mixins for claudefig TUI."""
 
-from .mixins import BackButtonMixin, FileInstanceMixin
+from .mixins import BackButtonMixin, FileInstanceMixin, ScrollNavigationMixin
 from .modal_screen import BaseModalScreen
 
 __all__ = [
     "BaseModalScreen",
     "BackButtonMixin",
     "FileInstanceMixin",
+    "ScrollNavigationMixin",
 ]
