+++ src/claudefig/tui/screens/overview.py
@@ -1,23 +1,26 @@
 """Project overview screen showing stats and quick actions."""
 
 from textual.app import ComposeResult
-from textual.containers import Container
+from textual.containers import Horizontal, VerticalScroll
+from textual.events import DescendantFocus, Key
 from textual.screen import Screen
 from textual.widgets import Button, Label, Static
 
 from claudefig.config import Config
 from claudefig.file_instance_manager import FileInstanceManager
 from claudefig.models import FileInstance, FileType
-from claudefig.tui.base import BackButtonMixin
+from claudefig.tui.base import BackButtonMixin, ScrollNavigationMixin
 from claudefig.tui.widgets import OverlayDropdown
 
 
-class OverviewScreen(Screen, BackButtonMixin):
+class OverviewScreen(Screen, BackButtonMixin, ScrollNavigationMixin):
     """Screen displaying project overview with stats and quick actions."""
 
     BINDINGS = [
         ("escape", "collapse_or_back", "Collapse/Back"),
         ("backspace", "collapse_or_back", "Collapse/Back"),
+        ("up", "focus_previous", "Focus Previous"),
+        ("down", "focus_next", "Focus Next"),
     ]
 
     def __init__(
@@ -38,7 +41,9 @@ class OverviewScreen(Screen, BackButtonMixin):
 
     def compose(self) -> ComposeResult:
         """Compose the overview screen."""
-        with Container(id="overview-screen"):
+        # can_focus=False prevents the container from being in the focus chain
+        # while still allowing it to be scrolled programmatically
+        with VerticalScroll(id="overview-screen", can_focus=False):
             # Title
             yield Label("PROJECT OVERVIEW", classes="screen-title")
